+++ xmake.lua
@@ -1,7 +1,7 @@
 includes("lib/commonlibf4")
 
 set_project("ESPExplorerAE")
-set_version("1.6.0")
+set_version("2.0.0")
 set_license("GPL-3.0")
 set_languages("c++23")
 set_warnings("allextra")